Move login requirement/redirect to its own hook
This splits up Authentication more and allows
possibility for un-authenticated routes
